# [3.10] bpo-45116: Fix inlining regressions on Windows Release build